+++ src/services/reducers/all-ingredients.jsx
@@ -1,0 +1,42 @@
+import {
+  GET_INGREDIENTS_REQUEST,
+  GET_INGREDIENTS_SUCCESS,
+  GET_INGREDIENTS_FAILED
+} from "../actions";
+
+
+const initialIngredients = {
+  allIngredients: [],
+  ingredientsRequest: false,
+  ingredientsFailed: false
+}
+
+
+export const allIngredientsReducer = (state = initialIngredients, action) => {
+  switch (action.type) {
+    case GET_INGREDIENTS_REQUEST: {
+      return {
+        ...state,
+        ingredientsRequest: true
+      };
+    }
+    case GET_INGREDIENTS_SUCCESS: {
+      return {
+        ...state,
+        ingredientsFailed: false,
+        allIngredients: action.data,
+        ingredientsRequest: false
+      };
+    }
+    case GET_INGREDIENTS_FAILED: {
+      return {
+        ...state,
+        ingredientsFailed: true,
+        ingredientsRequest: false
+      };
+    }
+    default: {
+      return state;
+    }
+  }
+}
