+++ src/utils/context.jsx
@@ -1,6 +1,3 @@
 import { createContext } from 'react';
 
-export const IngredientsContext = createContext([]);
-export const ConstructorContext = createContext([]);
-export const BunContext = createContext([]);
 export const OrderContext = createContext([]);
